tests/wpt/css-tests/css-flexbox-1_dev/html/display_inline-flex_exist.htm

<!DOCTYPE html>
<html><head>
    <title>CSS Flexible Box Test: display_inline-flex</title>
    <link href="http://www.intel.com" rel="author" title="Intel">
    <link href="http://www.w3.org/TR/css-flexbox-1/#flex-containers" rel="help">
    <meta content="dom" name="flags">
    <meta content="Check if the web engine can identify the display value inline-flex." name="assert">
    <style>
      div#test{
        display: inline-flex;
      }
    </style>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="test"></div>
    <script type="text/javascript">
        var t = async_test(document.title, {timeout:500});
        t.step(function () {
            assert_equals(window.getComputedStyle(document.querySelector('div#test')).display, "inline-flex", "Display value is");
        });
        t.done();
    </script>
  

</body></html>
